refactor: RepositoryModule을 Singleton과 ViewModel 스코프로 분리
